requirements.txt
flask==3.0.0
gunicorn==21.2.0
requests==2.32.3
msal==1.30.0
google-cloud-secret-manager==2.20.2
google-cloud-speech==2.25.0
google-generativeai==0.7.1
markdown2==2.4.13
google-cloud-aiplatform
google-cloud-speech
reportlab
weasyprint
cryptography==42.0.5